исправил баг с парсингом скорости ветра

+++ app/src/main/java/com/test/kabak/openweather/core/network/CurrentWeatherResponse.java
@@ -24,6 +24,6 @@ public class CurrentWeatherResponse {
 
     public static class WindObject {
         @SerializedName("speed")
-        public int windSpeed;
+        public float windSpeed;
     }
 }
